feat: Add commands for race management including setting barriers, checkpoints, start positions, and lobby locations

+++ surf-survival-events-events/surf-survival-event-race/src/main/kotlin/dev/slne/surf/survival/events/race/listener/QuitListener.kt
@@ -1,8 +1,12 @@
 package dev.slne.surf.survival.events.race.listener
 
+import com.github.shynixn.mccoroutine.folia.launch
 import dev.slne.surf.api.core.messages.adventure.sendText
+import dev.slne.surf.survival.events.race.listener.util.setOfflineLocation
+import dev.slne.surf.survival.events.race.plugin
 import dev.slne.surf.survival.events.race.service.RaceService
 import org.bukkit.Bukkit
+import org.bukkit.Location
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
 import org.bukkit.event.player.PlayerQuitEvent
@@ -11,11 +15,17 @@ object QuitListener : Listener {
     @EventHandler
     fun onQuit(event: PlayerQuitEvent) {
         val player = event.player
-        RaceService.removePlayer(player)
+        if (!RaceService.isInRace(player)) return
 
-        RaceService.getSpectatorPlayers().forEach {uuid ->
-            val playerSpectator = Bukkit.getPlayer(uuid)
-            playerSpectator?.sendText {
+        RaceService.removePlayer(player)
+        plugin.launch {
+            val offlinePlayer = Bukkit.getOfflinePlayer(player.uniqueId)
+            val location = Location(Bukkit.getWorld("world"), 0.5, 73.0, 0.5, 0f, 0f)
+            offlinePlayer.setOfflineLocation(location)
+        }
+
+        RaceService.getSpectatorPlayers().forEach { uuid ->
+            Bukkit.getPlayer(uuid)?.sendText {
                 appendInfoPrefix()
                 info("${player.name} hat das Rennen verlassen.")
             }
